improve --help.
Describe the -h and -v options.

Provide a better example of using a compound command. The previous
example, calling 'clear', was made obsolete by the recent addition
of the '-c' option, which clears the screen for you.

Provide a better example of ad-hoc GUI app created by using rerun
to connect executables together.