Simplify --verbose handling
The previous code never dropped filenames from the ever-expanding
list of files that had changed which it output on every new change.
The intent was that it would only display files that had changed
on that iteration. But it could never do this correctly since the
call to 'execute' started happening in th background.

So, I replaced it with this simpler code, that is a bit ugly and
only displays output for a moment before clearing the screen if
--clear is provided. But at least it's simple and doesn't break
in ridiculous ways.